refactor: Consolidate version update logic into a single script and remove obsolete files
